Refactor data views functions to views module
 - Refactor common functions to commons module
